perception testing & fine-tuning, camera & LiDAR config, Pixhawk frame check (changes? also compass seems to be messed up sometimes, need to check config for orientation and disable anything automatic there)
