+++ all_seaing_bringup/launch/vehicle.launch.py
@@ -22,6 +22,7 @@ def launch_setup(context, *args, **kwargs):
     bringup_prefix = get_package_share_directory("all_seaing_bringup")
     description_prefix = get_package_share_directory("all_seaing_description")
     driver_prefix = get_package_share_directory("all_seaing_driver")
+    perception_prefix = get_package_share_directory("all_seaing_perception")
 
     robot_urdf_file = os.path.join(
         description_prefix, "urdf", "fish_and_chips", "robot.urdf.xacro"
@@ -61,6 +62,7 @@ def launch_setup(context, *args, **kwargs):
         locations = yaml.safe_load(f)
 
     location = context.perform_substitution(LaunchConfiguration("location"))
+    use_slam = context.perform_substitution(LaunchConfiguration("use_slam"))
     comms = LaunchConfiguration("comms")
     use_bag = LaunchConfiguration("use_bag")
     is_indoors = str(locations[location]["indoors"]).lower()
@@ -102,8 +104,8 @@ def launch_setup(context, *args, **kwargs):
         ],
         parameters=[
             {"datum": [lat, lon, 0.0]},
-            {"yaw_offset": -np.pi/2.0},
-            {"odom_yaw_offset": -np.pi/2.0},
+            {"yaw_offset": np.pi/2.0},
+            {"odom_yaw_offset": np.pi/2.0},
             {"utm_zone": 17 if location == "nbpark" else 19}, # 19 for Boston, 17 for Florida
         ]
     )
@@ -295,271 +297,6 @@ def launch_setup(context, *args, **kwargs):
         ]
     )
 
-    grid_map_generator = launch_ros.actions.Node(
-        package="all_seaing_navigation",
-        executable="grid_map_generator.py",
-        parameters=[
-            {"global_frame_id": "map"},
-            {"timer_period": 1.0},
-            {"grid_dim": [800, 800]},
-            {"default_range": 60},
-            {"grid_resolution": 0.1},
-        ],
-    )
-
-    buoy_yolo_node = launch_ros.actions.Node(
-        package="all_seaing_perception",
-        executable="yolov8_node.py",
-        parameters=[
-            {"model": "roboboat_2025"},
-            {"label_config": "buoy_label_mappings"},
-            {"conf": 0.6},
-            {"use_color_names": False},
-        ],
-        remappings=[
-            ("image", "/zed/zed_node/rgb/image_rect_color"),
-            ("annotated_image", "annotated_image/buoy"),
-        ],
-        output="screen",
-    )
-
-    buoy_yolo_node_back_left = launch_ros.actions.Node(
-        package="all_seaing_perception",
-        executable="yolov8_node.py",
-        parameters=[
-            {"model": "roboboat_2025"},
-            {"label_config": "buoy_label_mappings"},
-            {"conf": 0.6},
-            {"use_color_names": False},
-        ],
-        remappings=[
-            ("image", "/back_left_oak/rgb/image_rect"),
-            ("annotated_image", "annotated_image/buoy/back_left"),
-            ("bounding_boxes", "bounding_boxes/back_left"),
-        ],
-        output="screen",
-    )
-
-    buoy_yolo_node_back_right = launch_ros.actions.Node(
-        package="all_seaing_perception",
-        executable="yolov8_node.py",
-        parameters=[
-            {"model": "roboboat_2025"},
-            {"label_config": "buoy_label_mappings"},
-            {"conf": 0.6},
-            {"use_color_names": False},
-        ],
-        remappings=[
-            ("image", "/back_right_oak/rgb/image_rect"),
-            ("annotated_image", "annotated_image/buoy/back_right"),
-            ("bounding_boxes", "bounding_boxes/back_right"),
-        ],
-        output="screen",
-    )
-
-    shape_yolo_node = launch_ros.actions.Node(
-        package="all_seaing_perception",
-        executable="yolov8_node.py",
-        parameters=[
-            {"model": "roboboat_shape_2025"},
-            {"label_config": "shape_label_mappings"},
-            {"conf": 0.4},
-            {"use_color_names": False},
-        ],
-        remappings=[
-            ("image", "turret_image"),
-            ("annotated_image", "annotated_image/shape"),
-            ("bounding_boxes", "shape_boxes"),
-        ],
-        output="screen",
-    )
-
-    static_shape_yolo_node = launch_ros.actions.Node(
-        package="all_seaing_perception",
-        executable="yolov8_node.py",
-        parameters=[
-            {"model": "roboboat_shape_2025"},
-            {"label_config": "shape_label_mappings"},
-            {"conf": 0.4},
-        ],
-        remappings=[
-            ("image", "/zed/zed_node/rgb/image_rect_color"),
-            ("annotated_image", "annotated_image/shape"),
-            ("bounding_boxes", "static_shape_boxes"),
-        ],
-        output="screen",
-    )
-
-    bbox_project_pcloud_node = launch_ros.actions.Node(
-        package="all_seaing_perception",
-        executable="bbox_project_pcloud",
-        output="screen",
-        remappings=[
-            ("camera_info_topic", "/zed/zed_node/rgb/camera_info"),
-            ("camera_topic", "/zed/zed_node/rgb/image_rect_color"),
-            ("lidar_topic", "/point_cloud/filtered")
-        ],
-        parameters=[
-            {"camera_name": "front"},
-            {"base_link_frame": "base_link"},
-            {"bbox_object_margin": 0.0},
-            {"color_label_mappings_file": inc_color_buoy_label_mappings},
-            {"obstacle_size_min": 2},
-            {"obstacle_size_max": 1000},
-            {"clustering_distance": 0.1},
-            {"matching_weights_file": matching_weights},
-            {"contour_matching_color_ranges_file": contour_matching_color_ranges},
-            {"is_sim": False},
-            {"label_list": True},
-        ]
-    )
-
-    bbox_project_pcloud_node_back_left = launch_ros.actions.Node(
-        package="all_seaing_perception",
-        executable="bbox_project_pcloud",
-        output="screen",
-        remappings=[
-            ("camera_info_topic", "/back_left_oak/rgb/camera_info"),
-            ("camera_topic", "/back_left_oak/rgb/image_rect"),
-            ("lidar_topic", "/point_cloud/filtered"),
-            ("bounding_boxes", "bounding_boxes/back_left"),
-        ],
-        parameters=[
-            {"camera_name": "back_left"},
-            {"base_link_frame": "base_link"},
-            {"bbox_object_margin": 0.0},
-            {"color_label_mappings_file": inc_color_buoy_label_mappings},
-            {"obstacle_size_min": 2},
-            {"obstacle_size_max": 1000},
-            {"clustering_distance": 0.1},
-            {"matching_weights_file": matching_weights},
-            {"contour_matching_color_ranges_file": contour_matching_color_ranges},
-            {"is_sim": False},
-            {"label_list": True},
-        ]
-    )
-
-    bbox_project_pcloud_node_back_right = launch_ros.actions.Node(
-        package="all_seaing_perception",
-        executable="bbox_project_pcloud",
-        output="screen",
-        remappings=[
-            ("camera_info_topic", "/back_right_oak/rgb/camera_info"),
-            ("camera_topic", "/back_right_oak/rgb/image_rect"),
-            ("lidar_topic", "/point_cloud/filtered"),
-            ("bounding_boxes", "bounding_boxes/back_right"),
-        ],
-        parameters=[
-            {"camera_name": "back_right"},
-            {"base_link_frame": "base_link"},
-            {"bbox_object_margin": 0.0},
-            {"color_label_mappings_file": inc_color_buoy_label_mappings},
-            {"obstacle_size_min": 2},
-            {"obstacle_size_max": 1000},
-            {"clustering_distance": 0.1},
-            {"matching_weights_file": matching_weights},
-            {"contour_matching_color_ranges_file": contour_matching_color_ranges},
-            {"is_sim": False},
-            {"label_list": True},
-        ]
-    )
-
-    multicam_detection_merge_node = launch_ros.actions.Node(
-        package="all_seaing_perception",
-        executable="multicam_detection_merge.py",
-        output="screen",
-        # arguments=['--ros-args', '--log-level', 'debug'],
-        remappings = [
-            ("detections/merged", "obstacle_map/local"),
-        ],
-        parameters = [
-            {"enable_front": True},
-            {"enable_back_left": True},
-            {"enable_back_right": True},
-            {"individual": False},
-            {"approximate": False},
-            {"delay": 0.1},
-        ]
-    )
-
-    obstacle_detector_raw_node = launch_ros.actions.Node(
-        package="all_seaing_perception",
-        executable="obstacle_detector",
-        remappings=[
-            ("point_cloud", "point_cloud/filtered"),
-        ],
-        parameters=[
-            {"base_link_frame": "base_link"},
-            {"global_frame_id": "map"},
-            {"clustering_distance": 1.0},
-            {"obstacle_size_min": 5},
-            {"range_max": 50.0},
-        ],
-    )
-
-    obstacle_detector_unlabeled_node = launch_ros.actions.Node(
-        package="all_seaing_perception",
-        executable="obstacle_detector",
-        remappings=[
-            ("point_cloud", "point_cloud/filtered"),
-            ("obstacle_map/raw", "obstacle_map/unlabeled")
-        ],
-        parameters=[
-            {"base_link_frame": "base_link"},
-            {"global_frame_id": "map"},
-            {"clustering_distance": 1.0},
-            {"obstacle_size_min": 5},
-            # {"obstacle_size_max": 300},
-            # {"obstacle_filter_pts_max": 100},
-            # {"obstacle_filter_area_max": 0.2},
-            {"obstacle_filter_length_max": 0.5},
-            # {"range_max": 50.0},
-        ],
-    )
-
-    object_tracking_map_node = launch_ros.actions.Node(
-        package="all_seaing_perception",
-        executable="object_tracking_map",
-        output="screen",
-        # arguments=['--ros-args', '--log-level', 'debug'],
-        remappings=[
-            ("detections", "obstacle_map/local"),
-            ("odometry/filtered", "odometry/gps"),
-        ],
-        parameters=[slam_params]
-    )
-
-    navigation_server = launch_ros.actions.Node(
-        package="all_seaing_navigation",
-        executable="navigation_server.py",
-        parameters=[
-            {"global_frame_id": "map"},
-            {"robot_frame_id": "base_link"},
-        ],
-        output="screen",
-    )
-
-    navigation_server_nomap = launch_ros.actions.Node(
-        package="all_seaing_navigation",
-        executable="navigation_server_nomap.py",
-        parameters=[
-            {"global_frame_id": "map"},
-            {"robot_frame_id": "base_link"},
-        ],
-        output="screen",
-    )
-
-    task_init_server = launch_ros.actions.Node(
-        package="all_seaing_autonomy",
-        executable="task_init.py",
-        parameters=[{"is_sim": False}],
-    )
-
-    delivery_server = launch_ros.actions.Node(
-        package="all_seaing_autonomy",
-        executable="delivery_server.py",
-    )
-
     central_hub = launch_ros.actions.Node(
         package="all_seaing_driver",
         executable="central_hub_ros.py",
@@ -584,7 +321,7 @@ def launch_setup(context, *args, **kwargs):
             ]
         ),
         launch_arguments={
-            "port": "/dev/ttyACM2"
+            "port": "/dev/ttyACM0"
         }.items(),
         condition=UnlessCondition(use_bag),
     )
@@ -651,7 +388,34 @@ def launch_setup(context, *args, **kwargs):
             ]),
         ),
     )
+
+    perception_ld = IncludeLaunchDescription(
+        PythonLaunchDescriptionSource(
+            [
+                bringup_prefix,
+                "/launch/perception.launch.py"
+            ]
+        ),
+        launch_arguments={
+            "location": location,
+            "use_slam": use_slam,
+            "use_bag": context.perform_substitution(use_bag),
+        }.items(),
+    )
     
+    tasks_ld = IncludeLaunchDescription(
+        PythonLaunchDescriptionSource(
+            [
+                bringup_prefix,
+                "/launch/tasks.launch.py"
+            ]
+        ),
+        launch_arguments={
+            "location": location,
+            "use_bag": context.perform_substitution(use_bag),
+        }.items(),
+    )
+
     return [
         control_mux,
         controller_node,
@@ -659,31 +423,11 @@ def launch_setup(context, *args, **kwargs):
         # ekf_node,
         # navsat_node,
         odometry_publisher_node,
-        # navigation_server,
-        # navigation_server_nomap,
-        obstacle_detector_node,
         point_cloud_filter_node,
         rover_custom_controller,
         rover_lora_controller,
         rviz_waypoint_sender,
         thrust_commander_node,
-        buoy_yolo_node,
-        buoy_yolo_node_back_left,
-        buoy_yolo_node_back_right,
-        # shape_yolo_node,
-        # static_shape_yolo_node,
-        bbox_project_pcloud_node,
-        bbox_project_pcloud_node_back_left,
-        bbox_project_pcloud_node_back_right,
-        multicam_detection_merge_node,
-        obstacle_detector_raw_node,
-        obstacle_detector_unlabeled_node,
-        # object_tracking_map_node,
-        # run_tasks,
-        # task_init_server, 
-        # follow_buoy_path,
-        # follow_buoy_pid,
-        # grid_map_generator,
         central_hub,
         # amcl_ld,
         # static_transforms_ld,
@@ -693,19 +437,22 @@ def launch_setup(context, *args, **kwargs):
         mavros_ld,
         zed_ld,
         oak_ld,
+        # perception_ld,
+        # tasks_ld,
     ]
 
 
 def generate_launch_description():
     return LaunchDescription(
         [
-            DeclareLaunchArgument("location", default_value="pavillion"),
+            DeclareLaunchArgument("location", default_value="boathouse"),
             DeclareLaunchArgument(
-                "comms", default_value="wifi", choices=["wifi", "lora", "custom"]
+                "comms", default_value="custom", choices=["wifi", "lora", "custom"]
             ),
             DeclareLaunchArgument(
                 "use_bag", default_value="false", choices=["true", "false"]
             ),
+            DeclareLaunchArgument("use_slam", default_value="true", choices=["true", "false"]),
             OpaqueFunction(function=launch_setup),
         ]
     )
