+++ all_seaing_bringup/rosbag_record/record_slam.bash
@@ -4,4 +4,4 @@ current_date_time=$(date +"%m_%d__%H_%M")
 file_name="slam_$current_date_time"
 echo "Recording lidar & camera & detections & localization & maps to $file_name"
 # export ROS_DOMAIN_ID=10
-ros2 bag record -o ../rosbags/$file_name /zed/zed_node/rgb/image_rect_color /zed/zed_node/rgb/camera_info /back_left_oak/rgb/image_rect /back_left_oak/rgb/camera_info /back_right_oak/rgb/image_rect /back_right_oak/rgb/camera_info /bounding_boxes /bounding_boxes/back_left /bounding_boxes/back_right /shape_boxes /static_shape_boxes /image/segmented /bounding_boxes_ycrcb /image/segmented_ycrcb /annotated_image/buoy /annotated_image/buoy/back_left /annotated_image/buoy/back_right /velodyne_points /point_cloud/filtered /labeled_object_point_clouds/front /detections/front /detections/merged /refined_object_point_clouds_viz/front /refined_object_segments_viz/front /detections/back_left /refined_object_point_clouds_viz/back_left /refined_object_segments_viz/back_left /detections/back_right /refined_object_point_clouds_viz/back_right /refined_object_segments_viz/back_right /mavros/global_position/global /mavros/imu/data /mavros/imu/data_raw /mavros/imu/mag /mavros/local_position/local /mavros/local_position/pose /mavros/local_position/odom /mavros/global_position/raw/fix /odometry/filtered /odometry/gps /tf /tf_static /obstacle_map/local /obstacle_map/global /obstacle_map/raw /obstacle_map/unlabeled
+ros2 bag record -o ../rosbags/$file_name /zed/zed_node/rgb/image_rect_color /zed/zed_node/rgb/camera_info /back_left_oak/rgb/image_rect/compressed /back_left_oak/rgb/camera_info /back_right_oak/rgb/image_rect/compressed /back_right_oak/rgb/camera_info /bounding_boxes /bounding_boxes/back_left /bounding_boxes/back_right /shape_boxes /static_shape_boxes /image/segmented /bounding_boxes_ycrcb /image/segmented_ycrcb /annotated_image/buoy /annotated_image/buoy/back_left /annotated_image/buoy/back_right /velodyne_points /point_cloud/filtered /labeled_object_point_clouds/front /detections/front /detections/merged /refined_object_point_clouds_viz/front /refined_object_segments_viz/front /detections/back_left /refined_object_point_clouds_viz/back_left /refined_object_segments_viz/back_left /detections/back_right /refined_object_point_clouds_viz/back_right /refined_object_segments_viz/back_right /mavros/global_position/global /mavros/imu/data /mavros/imu/data_raw /mavros/imu/mag /mavros/local_position/local /mavros/local_position/pose /mavros/local_position/odom /mavros/global_position/raw/fix /odometry/filtered /odometry/gps /tf /tf_static /obstacle_map/local /obstacle_map/global /obstacle_map/raw /obstacle_map/unlabeled
