+++ all_seaing_bringup/scripts/on_boot.bash
@@ -1,0 +1,3 @@
+#!/bin/bash
+chmod 666 /dev/ttyACM*
+source /opt/ros/humble/setup.bash
