+++ all_seaing_perception/all_seaing_perception/yolov8_node.py
@@ -44,6 +44,8 @@ class Yolov8Node(Node):
 
         if self.device == "default":
             self.device = "cuda" if torch.cuda.is_available() else "cpu"
+        
+        self.get_logger().info(f'YOLO DEVICE USED: {self.device}')
 
         label_config = self.get_parameter("label_config").get_parameter_value().string_value
         yaml_file_path = os.path.join(bringup_prefix, "config","perception", label_config + ".yaml")
